// apps/sot-backend/service-order/src/lei/vehicles/dto/create-vehicle.dto.ts
export class CreateVehicleDto {
  automaker: string;
  project: string;
  responsible: string;
  model: string;
  color: string;
  chassis: string;
  fleet: string;
  comments: string;
  image: string;
  nf_number: string;
  nf_emission_date: string;
  loan_expiration_date: string;
  nf_archive_location: string;
}